requirements.txt
streamlit
openai
chromadb
pypdf
python-dotenv
tiktoken
pytest